requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.5.0
numpy>=1.20.0
Pillow>=8.0.0
pycocotools>=2.0.0
tqdm>=4.60.0
